Fixed testReading

+++ tests/testReading.c
@@ -39,7 +39,7 @@ int main() {
 
   /* Properties of the ICs */
   const double boxSize = 1.;
-  const int L = 4;
+  const size_t L = 4;
   const double rho = 2.;
 
   /* Read data */
